feat(core): enforcement layer + org_tree + budget tracker hardening
- enforcement/token_capture.py (NEW): captures Anthropic token usage from runtime for budget tracker reconciliation
- core/org_tree.py (NEW): in-memory squad-skill org graph + traversal helpers (powers cognitive dashboard)
- core/runtime.py: token_capture hook + minor runtime instrumentation
- core/install_service.py: tightened version detection for installer compatibility
- schemas/contracts.py: schema fix for budget contract validation
- finance/budget_tracker.py: +78 lines — per-client aggregation hooks + percentage signal

Part of v12.5.0 prep.

+++ orchestrator/schemas/contracts.py
@@ -214,7 +214,7 @@ class HealthResponse(BaseModel):
     status: str
     state: str = "?"
     autonomy: str = "?"
-    health: int = 0
+    health: float = 0.0
     timestamp: str
